requirements.txt
# ── Web framework ─────────────────────────────
flask==3.0.3
flask-cors==4.0.1

# ── Deep learning (CPU build — change to torch with CUDA if you have a GPU)
torch==2.9.0
torchvision==0.24.0

# ── Image processing ──────────────────────────
Pillow==11.2.1
opencv-python-headless==4.9.0.80
numpy==1.26.4

# ── Production server ─────────────────────────
gunicorn==22.0.0

# ── Utilities ─────────────────────────────────
python-dotenv==1.0.1